Added Table Writer calls to Perfect Foresight Runs
PF runs now properly call table_writer to register the costs in OutputCost
