+++ temoa/temoa_model/run_actions.py
@@ -39,6 +39,7 @@ from pyomo.environ import DataPortal, Suffix, Var, Constraint, value, UnknownSol
 from pyomo.opt import SolverResults, SolverStatus, TerminationCondition
 
 from temoa.temoa_model.pformat_results import pformat_results
+from temoa.temoa_model.table_writer import TableWriter
 from temoa.temoa_model.temoa_config import TemoaConfig
 from temoa.temoa_model.temoa_model import TemoaModel
 
@@ -303,6 +304,9 @@ def handle_results(instance: TemoaModel, results, options: TemoaConfig):
         SE.flush()
 
     output_stream = pformat_results(instance, results, options)
+    table_writer = TableWriter(config=options)
+    table_writer.clear_scenario()
+    table_writer.write_costs(M=instance)
 
     if not options.silent:
         SE.write('\r[%8.2f] Results processed.\n' % (time() - hack))
